pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Chengwei Ye", email = "chengweiye18@gmail.com"},
]
dependencies = [
    "markdown>=3.3.7",
    "python-wordpress-xmlrpc>=2.3",
    "python-frontmatter>=1.0.0",
]
requires-python = ">=3.8"
license = {text = "GNU GENERAL PUBLIC LICENSE"}

[tool.pdm]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
